# Optimize while loop to use 4 elements instead of 7 for calculation, improving performance on older CPUs