Description:
I investigated and optimized the function in a separate environment. The standard function, when processing an array with size 640 MB + 4 bytes, took approximately 0.562 seconds to execute. After rearranging and optimizing the code, the execution time was reduced to approximately 0.247 seconds. This represents a significant performance improvement.

Testing:

Environment: Tested on a machine with an Intel Core i7-6700HQ CPU.
Data: Used 640 MB + 4 byte
Results:
Original function execution time: ~0.562 seconds
Optimized function execution time: ~0.247 seconds
Impact:
This optimization should improve the performance of the function, especially on older CPUs or in scenarios with large datasets.