@sundb 2026-02-10
we can close this one via https://github.com/redis/redis/pull/14309,  Intel Core i7-6700HQ also supports these instructions.
@retsetB feel free to reopen it if you think it's still worth it.